+++ nest-reader-dem/src/main/java/org/esa/nest/dataio/dem/BaseElevationModel.java
@@ -34,13 +34,13 @@ public abstract class BaseElevationModel implements ElevationModel, Resampling.R
     protected final int NUM_X_TILES;
     protected final int NUM_Y_TILES;
     protected final int NUM_PIXELS_PER_TILE;
-    private final float NUM_PIXELS_PER_TILEinv;
+    private final double NUM_PIXELS_PER_TILEinv;
     protected final float NO_DATA_VALUE;
     protected final int DEGREE_RES;
     protected final int RASTER_WIDTH;
     protected final int RASTER_HEIGHT;
-    protected final float DEGREE_RES_BY_NUM_PIXELS_PER_TILE;
-    protected final float DEGREE_RES_BY_NUM_PIXELS_PER_TILEinv;
+    protected final double DEGREE_RES_BY_NUM_PIXELS_PER_TILE;
+    protected final double DEGREE_RES_BY_NUM_PIXELS_PER_TILEinv;
 
     protected final ElevationModelDescriptor descriptor;
     private final ElevationFile[][] elevationFiles;
@@ -61,14 +61,14 @@ public abstract class BaseElevationModel implements ElevationModel, Resampling.R
         NUM_Y_TILES = descriptor.getNumYTiles();
         NO_DATA_VALUE = descriptor.getNoDataValue();
         NUM_PIXELS_PER_TILE = descriptor.getPixelRes();
-        NUM_PIXELS_PER_TILEinv = 1.0f / (float)NUM_PIXELS_PER_TILE;
+        NUM_PIXELS_PER_TILEinv = 1.0 / (double)NUM_PIXELS_PER_TILE;
         DEGREE_RES = descriptor.getDegreeRes();
 
         RASTER_WIDTH = NUM_X_TILES * NUM_PIXELS_PER_TILE;
         RASTER_HEIGHT = NUM_Y_TILES * NUM_PIXELS_PER_TILE;
 
-        DEGREE_RES_BY_NUM_PIXELS_PER_TILE = DEGREE_RES / (float)NUM_PIXELS_PER_TILE;
-        DEGREE_RES_BY_NUM_PIXELS_PER_TILEinv = 1.0f / DEGREE_RES_BY_NUM_PIXELS_PER_TILE;
+        DEGREE_RES_BY_NUM_PIXELS_PER_TILE = DEGREE_RES / (double)NUM_PIXELS_PER_TILE;
+        DEGREE_RES_BY_NUM_PIXELS_PER_TILEinv = 1.0 / DEGREE_RES_BY_NUM_PIXELS_PER_TILE;
 
         elevationFiles = createElevationFiles();    // must be last
     }
@@ -86,7 +86,7 @@ public abstract class BaseElevationModel implements ElevationModel, Resampling.R
     }
 
     public final synchronized float getElevation(final GeoPos geoPos) throws Exception {
-        final float pixelY = getIndexY(geoPos);
+        final double pixelY = getIndexY(geoPos);
         if (pixelY < 0) {
             return NO_DATA_VALUE;
         }
@@ -100,14 +100,14 @@ public abstract class BaseElevationModel implements ElevationModel, Resampling.R
         return elevation;
     }
 
-    public abstract float getIndexX(final GeoPos geoPos);
+    public abstract double getIndexX(final GeoPos geoPos);
 
-    public abstract float getIndexY(final GeoPos geoPos);
+    public abstract double getIndexY(final GeoPos geoPos);
 
     public abstract GeoPos getGeoPos(final PixelPos pixelPos);
 
     public PixelPos getIndex(final GeoPos geoPos) {
-        return new PixelPos(getIndexX(geoPos), getIndexY(geoPos));
+        return new PixelPos((float)getIndexX(geoPos), (float)getIndexY(geoPos));
     }
 
     public void dispose() {
