+++ nest-reader-dem/src/main/java/org/esa/nest/dataio/dem/ace/ACEElevationModel.java
@@ -34,19 +34,19 @@ public class ACEElevationModel extends BaseElevationModel {
     }
 
     @Override
-    public float getIndexX(final GeoPos geoPos) {
-        return (geoPos.lon + 180.0f) / DEGREE_RES_BY_NUM_PIXELS_PER_TILE;
+    public double getIndexX(final GeoPos geoPos) {
+        return (geoPos.lon + 180.0) / DEGREE_RES_BY_NUM_PIXELS_PER_TILE;
     }
 
     @Override
-    public float getIndexY(final GeoPos geoPos) {
-        return RASTER_HEIGHT - (geoPos.lat + 90.0f) / DEGREE_RES_BY_NUM_PIXELS_PER_TILE;
+    public double getIndexY(final GeoPos geoPos) {
+        return RASTER_HEIGHT - (geoPos.lat + 90.0) / DEGREE_RES_BY_NUM_PIXELS_PER_TILE;
     }
 
     @Override
     public GeoPos getGeoPos(PixelPos pixelPos) {
-        float pixelLat = (RASTER_HEIGHT - pixelPos.y) / DEGREE_RES_BY_NUM_PIXELS_PER_TILE - 90.0f;
-        float pixelLon = pixelPos.x / DEGREE_RES_BY_NUM_PIXELS_PER_TILE - 180.0f;
+        float pixelLat = (float)((RASTER_HEIGHT - pixelPos.y) / DEGREE_RES_BY_NUM_PIXELS_PER_TILE - 90.0);
+        float pixelLon = (float)(pixelPos.x / DEGREE_RES_BY_NUM_PIXELS_PER_TILE - 180.0);
         return new GeoPos(pixelLat, pixelLon);
     }
